pyproject.toml
[project]
name = "her-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-core>=1.30.0",
    "azure-identity>=1.15.0",
    "click>=8.1.8",
    "elevenlabs>=0.2.26",
    "fastapi>=0.89.0",
    "graphiti-core>=0.1.0",
    "modal>=0.54.3",
    "neo4j>=5.0.0",
    "openai>=1.6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "uvicorn>=0.24.0",
    "zep-cloud>=2.16.0",
]
